Add implementation of command to get organisation by id
