+++ src/SFA.DAS.ReferenceData.Api.Client/IReferenceDataApiClient.cs
@@ -1,8 +1,8 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 using SFA.DAS.ReferenceData.Api.Client.Dto;
-using SFA.DAS.ReferenceData.Api.Client.Exceptions;
 using SFA.DAS.ReferenceData.Types;
+using SFA.DAS.ReferenceData.Types.Exceptions;
 
 namespace SFA.DAS.ReferenceData.Api.Client
 {
