+++ src/SFA.DAS.ReferenceData.Api.Client/SFA.DAS.ReferenceData.Api.Client.csproj
@@ -57,11 +57,8 @@
     <Compile Include="Dto\PagedApiResponse.cs" />
     <Compile Include="IReferenceDataApiClient.cs" />
     <Compile Include="IReferenceDataApiConfiguration.cs" />
-    <Compile Include="Exceptions\OperationNotSupportedForOrganisationType.cs" />
-    <Compile Include="Exceptions\OrganisationNotFoundExeption.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ReferenceDataApiClient.cs" />
-    <Compile Include="Exceptions\ReferenceDataException.cs" />
     <Compile Include="SecureHttpClient.cs" />
   </ItemGroup>
   <ItemGroup>
